+++ python/docstrings.cpp
@@ -2264,6 +2264,46 @@ Notes:
     for the filtering to be executed.  If either one is ``None`` they
     are both ignored.)!");
 
+PyDoc_STRVAR(cell_object_get_paths_doc, R"!(get_paths(apply_repetitions=True, depth=None, layer=None, datatype=None) -> list
+
+Return a copy of all paths in the cell.
+
+Args:
+    apply_repetitions: Define whether repetitions should be applied in
+      the created paths.
+    depth: If non negative, indicates the number of reference levels
+      processed recursively.  A value of 0 will result in no references
+      being visited.
+    layer: If set, only paths in the defined layer and data type are
+      returned.
+    datatype: If set, only paths in the defined layer and data type are
+      returned.
+
+Notes:
+    Arguments ``layer`` and ``datatype`` must both be set to integers
+    for the filtering to be executed.  If either one is ``None`` they
+    are both ignored.)!");
+
+PyDoc_STRVAR(cell_object_get_labels_doc, R"!(get_labels(apply_repetitions=True, depth=None, layer=None, texttype=None) -> list
+
+Return a copy of all labels in the cell.
+
+Args:
+    apply_repetitions: Define whether repetitions should be applied in
+      the created labels.
+    depth: If non negative, indicates the number of reference levels
+      processed recursively.  A value of 0 will result in no references
+      being visited.
+    layer: If set, only labels in the defined layer and text type are
+      returned.
+    texttype: If set, only labels in the defined layer and text type
+      are returned.
+
+Notes:
+    Arguments ``layer`` and ``texttype`` must both be set to integers
+    for the filtering to be executed.  If either one is ``None`` they
+    are both ignored.)!");
+
 PyDoc_STRVAR(cell_object_flatten_doc, R"!(flatten(apply_repetitions=True) -> self
 
 Transform all references into polygons, paths and labels.