+++ src/reference.h
@@ -87,11 +87,14 @@ struct Reference {
     // polygonal representation of paths are also included in polygons.  If
     // filter is true, only polygons in the indicated layer and data type are
     // created.
-    void polygons(bool apply_repetitions, bool include_paths, int64_t depth,
-                  bool filter, uint32_t layer, uint32_t datatype, Array<Polygon*>& result) const;
-    void flexpaths(bool apply_repetitions, int64_t depth, Array<FlexPath*>& result) const;
-    void robustpaths(bool apply_repetitions, int64_t depth, Array<RobustPath*>& result) const;
-    void labels(bool apply_repetitions, int64_t depth, Array<Label*>& result) const;
+    void polygons(bool apply_repetitions, bool include_paths, int64_t depth, bool filter,
+                  uint32_t layer, uint32_t datatype, Array<Polygon*>& result) const;
+    void flexpaths(bool apply_repetitions, int64_t depth, bool filter, uint32_t layer,
+                   uint32_t datatype, Array<FlexPath*>& result) const;
+    void robustpaths(bool apply_repetitions, int64_t depth, bool filter, uint32_t layer,
+                     uint32_t datatype, Array<RobustPath*>& result) const;
+    void labels(bool apply_repetitions, int64_t depth, bool filter, uint32_t layer,
+                uint32_t texttype, Array<Label*>& result) const;
 
     // These functions output the reference in the GDSII and SVG formats.  They
     // are not supposed to be called by the user.